// src/app/models/chiefcomplaintdetail.model.ts
export class ChiefComplaintDetail { 
    chiefComplaintId?: number;
    chiefComplaintCode: string;
    chiefComplaintDescription: string;
    code: string;
    description: string;
    id: number;
    createdBy: string;
    createdDateTime: string;
    modifiedBy: string;
    modifiedDateTime: string;
    status: boolean;
}